if statement for torch and fixes
